More control of files including gitid and giturl
The current solution allows the user to add gitid and giturl
to the control file, .dsc in source, and debian/control in
binary. In some cases this is not applicable, example you
may want to keep the URL hidden from distributed binaries.

The new solution removes defaults for these commands and
now adds the options [None, 'both', 'source', 'binary']

None = Not applied to any files
both = Applied to both source and binary files
source = Applied to source files only
binary = Applied to binary files only

Additionally the options can be set in a configuration file.